The `lastInRange` value is redundant to `lastValue`, and by removing it I was able to see significant performance improvements when running the `bench_test.go`:

<details>
  <summary>


# Yrate Old
</summary>

```
BenchmarkRangeQuery/expr=yrate(a_one[1h]),steps=1-12        	   56229	     21123 ns/op	   25152 B/op	     187 allocs/op
BenchmarkRangeQuery/expr=yrate(a_one[1h]),steps=10-12       	   45560	     25054 ns/op	   25154 B/op	     187 allocs/op
BenchmarkRangeQuery/expr=yrate(a_one[1h]),steps=100-12      	   18744	     63534 ns/op	   25728 B/op	     192 allocs/op
BenchmarkRangeQuery/expr=yrate(a_one[1h]),steps=1000-12     	    2733	    459238 ns/op	   28650 B/op	     231 allocs/op
BenchmarkRangeQuery/expr=yrate(a_ten[1h]),steps=1-12        	    9026	    130511 ns/op	   42722 B/op	     471 allocs/op
BenchmarkRangeQuery/expr=yrate(a_ten[1h]),steps=10-12       	    7056	    168845 ns/op	   42700 B/op	     471 allocs/op
BenchmarkRangeQuery/expr=yrate(a_ten[1h]),steps=100-12      	    2144	    573474 ns/op	   45420 B/op	     512 allocs/op
BenchmarkRangeQuery/expr=yrate(a_ten[1h]),steps=1000-12     	     261	   4347225 ns/op	   66361 B/op	     851 allocs/op
BenchmarkRangeQuery/expr=yrate(a_hundred[1h]),steps=1-12    	    1020	   1167475 ns/op	  216727 B/op	    3270 allocs/op
BenchmarkRangeQuery/expr=yrate(a_hundred[1h]),steps=10-12   	     765	   1616422 ns/op	  216600 B/op	    3270 allocs/op
BenchmarkRangeQuery/expr=yrate(a_hundred[1h]),steps=100-12  	     210	   5657878 ns/op	  240746 B/op	    3671 allocs/op
BenchmarkRangeQuery/expr=yrate(a_hundred[1h]),steps=1000-12 	      25	  46485908 ns/op	  434446 B/op	    6873 allocs/op
```

</details>

<details>
<summary>

# Yrate without lastInRange variable
  </summary>

```
BenchmarkRangeQuery/expr=yrate(a_one[1h]),steps=1-12        	   56678	     20917 ns/op	   25156 B/op	     187 allocs/op
BenchmarkRangeQuery/expr=yrate(a_one[1h]),steps=10-12       	   48253	     24430 ns/op	   25149 B/op	     187 allocs/op
BenchmarkRangeQuery/expr=yrate(a_one[1h]),steps=100-12      	   20371	     58235 ns/op	   25692 B/op	     192 allocs/op
BenchmarkRangeQuery/expr=yrate(a_one[1h]),steps=1000-12     	    3051	    394467 ns/op	   28660 B/op	     231 allocs/op
BenchmarkRangeQuery/expr=yrate(a_ten[1h]),steps=1-12        	    9283	    124002 ns/op	   42731 B/op	     471 allocs/op
BenchmarkRangeQuery/expr=yrate(a_ten[1h]),steps=10-12       	    7370	    157232 ns/op	   42739 B/op	     471 allocs/op
BenchmarkRangeQuery/expr=yrate(a_ten[1h]),steps=100-12      	    2391	    494936 ns/op	   45439 B/op	     512 allocs/op
BenchmarkRangeQuery/expr=yrate(a_ten[1h]),steps=1000-12     	     313	   3811625 ns/op	   66358 B/op	     851 allocs/op
BenchmarkRangeQuery/expr=yrate(a_hundred[1h]),steps=1-12    	    1017	   1144133 ns/op	  216587 B/op	    3270 allocs/op
BenchmarkRangeQuery/expr=yrate(a_hundred[1h]),steps=10-12   	     783	   1478840 ns/op	  216688 B/op	    3270 allocs/op
BenchmarkRangeQuery/expr=yrate(a_hundred[1h]),steps=100-12  	     244	   4875852 ns/op	  241273 B/op	    3671 allocs/op
BenchmarkRangeQuery/expr=yrate(a_hundred[1h]),steps=1000-12 	      30	  38099921 ns/op	  433979 B/op	    6873 allocs/op
```
</details>